safari-600.7-branch/LayoutTests/fast/dom/shadow/get-element-by-id-in-shadow-root-expected.txt

Tests to ensure that ShadowRoot.getElementById can get an element in ShadowRoot by its id. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getNodeInShadowTreeStack('divA/inputB').id is "inputB"
PASS getNodeInShadowTreeStack('divA/divC').id is "divC"
PASS getNodeInShadowTreeStack('divA/divC/inputD').id is "inputD"
PASS getNodeInShadowTreeStack('divA/divC//inputE').id is "inputE"
PASS getNodeInShadowTreeStack('divA/divC///inputF').id is "inputF"
PASS successfullyParsed is true

TEST COMPLETE
